refactor StepResult to results/ dir to support more results related objects in future
