+++ src/ploigos_step_runner/__init__.py
@@ -792,9 +792,8 @@ Example Running the 'generate-metadata' step
 """
 
 import __main__
-
 from ploigos_step_runner.exceptions import StepRunnerException
+from ploigos_step_runner.results import StepResult
 from ploigos_step_runner.step_implementer import DefaultSteps, StepImplementer
-from ploigos_step_runner.step_result import StepResult
 from ploigos_step_runner.step_runner import StepRunner
 from ploigos_step_runner.workflow_result import WorkflowResult
