+++ src/ploigos_step_runner/results/__init__.py
@@ -1,0 +1,4 @@
+"""Results for Ploigos workflow.
+"""
+
+from ploigos_step_runner.results.step_result import StepResult
