+++ src/ploigos_step_runner/step_implementer.py
@@ -10,7 +10,7 @@ from contextlib import redirect_stderr, redirect_stdout
 from pathlib import Path
 
 from ploigos_step_runner.config.config_value import ConfigValue
-from ploigos_step_runner.step_result import StepResult
+from ploigos_step_runner import StepResult
 from ploigos_step_runner.utils.io import TextIOIndenter
