+++ src/ploigos_step_runner/step_implementers/sign_container_image/curl_push.py
@@ -50,7 +50,7 @@ from io import StringIO
 import sh
 from ploigos_step_runner import StepImplementer
 from ploigos_step_runner.exceptions import StepRunnerException
-from ploigos_step_runner.step_result import StepResult
+from ploigos_step_runner import StepResult
 from ploigos_step_runner.utils.io import create_sh_redirect_to_multiple_streams_fn_callback
 
 DEFAULT_CONFIG = {
