+++ src/ploigos_step_runner/step_implementers/static_code_analysis/sonarqube.py
@@ -89,7 +89,7 @@ import sys
 import sh
 from ploigos_step_runner import StepImplementer
 from ploigos_step_runner.exceptions import StepRunnerException
-from ploigos_step_runner.step_result import StepResult
+from ploigos_step_runner import StepResult
 
 DEFAULT_CONFIG = {
     'properties': './sonar-project.properties',
