+++ src/ploigos_step_runner/step_implementers/tag_source/git.py
@@ -45,7 +45,7 @@ from io import StringIO
 import sh
 from ploigos_step_runner import StepImplementer
 from ploigos_step_runner.exceptions import StepRunnerException
-from ploigos_step_runner.step_result import StepResult
+from ploigos_step_runner import StepResult
 
 DEFAULT_CONFIG = {}
