+++ src/ploigos_step_runner/step_implementers/uat/maven_selenium_cucumber.py
@@ -45,7 +45,7 @@ import sh
 from ploigos_step_runner.config.config_value import ConfigValue
 from ploigos_step_runner.exceptions import StepRunnerException
 from ploigos_step_runner.step_implementers.shared.maven_generic import MavenGeneric
-from ploigos_step_runner.step_result import StepResult
+from ploigos_step_runner import StepResult
 from ploigos_step_runner.utils.io import create_sh_redirect_to_multiple_streams_fn_callback
 
 DEFAULT_CONFIG = {
