+++ src/ploigos_step_runner/step_implementers/validate_environment_configuration/configlint.py
@@ -73,7 +73,7 @@ import sys
 
 import sh
 from ploigos_step_runner import StepImplementer
-from ploigos_step_runner.step_result import StepResult
+from ploigos_step_runner import StepResult
 from ploigos_step_runner.utils.io import create_sh_redirect_to_multiple_streams_fn_callback
 
 DEFAULT_CONFIG = {
