+++ tests/helpers/base_step_implementer_test_case.py
@@ -9,7 +9,7 @@ import yaml
 from ploigos_step_runner.step_runner import StepRunner
 from ploigos_step_runner.config.config import Config
 from ploigos_step_runner.exceptions import StepRunnerException
-from ploigos_step_runner.step_result import StepResult
+from ploigos_step_runner import StepResult
 from ploigos_step_runner.workflow_result import WorkflowResult
 
 from .base_test_case import BaseTestCase
