+++ tests/step_implementers/generate_metadata/test_maven_generate_metadata.py
@@ -7,7 +7,7 @@ from testfixtures import TempDirectory
 from tests.helpers.base_step_implementer_test_case import \
     BaseStepImplementerTestCase
 from ploigos_step_runner.step_implementers.generate_metadata import Maven
-from ploigos_step_runner.step_result import StepResult
+from ploigos_step_runner import StepResult
 
 
 class TestStepImplementerMavenGenerateMetadata(BaseStepImplementerTestCase):
