+++ tests/step_implementers/package/test_maven_package.py
@@ -7,11 +7,11 @@ from pathlib import Path
 from unittest.mock import patch
 
 import sh
+from ploigos_step_runner import StepResult
+from ploigos_step_runner.step_implementers.package import Maven
 from testfixtures import TempDirectory
 from tests.helpers.base_step_implementer_test_case import \
     BaseStepImplementerTestCase
-from ploigos_step_runner import StepResult
-from ploigos_step_runner.step_implementers.package import Maven
 
 
 class TestStepImplementerMavenPackageBase(BaseStepImplementerTestCase):
