+++ tests/step_implementers/push_container_image/test_skopeo_push_container_image.py
@@ -13,7 +13,7 @@ from tests.helpers.base_step_implementer_test_case import \
     BaseStepImplementerTestCase
 from tests.helpers.test_utils import Any
 from ploigos_step_runner.step_implementers.push_container_image import Skopeo
-from ploigos_step_runner.step_result import StepResult
+from ploigos_step_runner import StepResult
 
 
 class TestStepImplementerSkopeoSourceBase(BaseStepImplementerTestCase):
