+++ tests/step_implementers/sign_container_image/test_podman_sign.py
@@ -14,7 +14,7 @@ from tests.helpers.base_step_implementer_test_case import \
 from tests.helpers.test_utils import Any, StringRegexParam
 from ploigos_step_runner.exceptions import StepRunnerException
 from ploigos_step_runner.step_implementers.sign_container_image import PodmanSign
-from ploigos_step_runner.step_result import StepResult
+from ploigos_step_runner import StepResult
 
 
 class TestStepImplementerSignContainerImagePodman(BaseStepImplementerTestCase):
