+++ tests/step_implementers/static_code_analysis/test_sonarqube_static_code_analysis.py
@@ -7,11 +7,12 @@ import sys
 from unittest.mock import patch
 
 import sh
+from ploigos_step_runner import StepResult, StepRunnerException
+from ploigos_step_runner.step_implementers.static_code_analysis import \
+    SonarQube
 from testfixtures import TempDirectory
 from tests.helpers.base_step_implementer_test_case import \
     BaseStepImplementerTestCase
-from ploigos_step_runner import StepResult, StepRunnerException
-from ploigos_step_runner.step_implementers.static_code_analysis import SonarQube
 
 
 class TestStepImplementerSonarQubePackageBase(BaseStepImplementerTestCase):
