+++ tests/step_implementers/uat/test_maven_selenium_cucumber.py
@@ -11,7 +11,7 @@ from tests.helpers.maven_step_implementer_test_case import \
 from tests.helpers.test_utils import Any
 from ploigos_step_runner.exceptions import StepRunnerException
 from ploigos_step_runner.step_implementers.uat import MavenSeleniumCucumber
-from ploigos_step_runner.step_result import StepResult
+from ploigos_step_runner import StepResult
 from ploigos_step_runner.utils.file import create_parent_dir
