+++ tests/step_implementers/unit_test/test_maven_unit_test.py
@@ -10,7 +10,7 @@ from tests.helpers.maven_step_implementer_test_case import \
 from tests.helpers.test_utils import Any
 from ploigos_step_runner.config.config import Config
 from ploigos_step_runner.step_implementers.unit_test import Maven
-from ploigos_step_runner.step_result import StepResult
+from ploigos_step_runner import StepResult
 from ploigos_step_runner.utils.file import create_parent_dir
 from ploigos_step_runner.workflow_result import WorkflowResult
