+++ tests/step_implementers/validate_environment_configuration/test_configlint_from_argocd.py
@@ -8,7 +8,7 @@ from tests.helpers.base_step_implementer_test_case import \
     BaseStepImplementerTestCase
 from ploigos_step_runner.step_implementers.validate_environment_configuration import \
     ConfiglintFromArgocd
-from ploigos_step_runner.step_result import StepResult
+from ploigos_step_runner import StepResult
 
 
 class TestStepImplementerConfiglintFromArgocd(BaseStepImplementerTestCase):
