+++ tests/test_workflow_result.py
@@ -7,7 +7,7 @@ from testfixtures import TempDirectory
 
 from tests.helpers.base_test_case import BaseTestCase
 
-from ploigos_step_runner.step_result import StepResult
+from ploigos_step_runner import StepResult
 from ploigos_step_runner.workflow_result import WorkflowResult
 from ploigos_step_runner.exceptions import StepRunnerException
